requirements.txt
streamlit
gTTS
Pillow